Refactor (APIMF-3491): extracted link include property parser
